refactor: extract GameRenderer from game.ts
Moves the rendering block (render() plus renderTraversalTunnel,
renderCosmicRebirth, getDestinationPreviewObjects,
renderChunkBoundaries, and the culling helpers) into a new
GameRenderer class in src/graphics/.

The code moved verbatim via accessor mirrors over the game''s component
graph, with one deliberate exception: render() now looks up destination
previews through the Game facade so existing test spies on
game.getDestinationPreviewObjects still intercept the call.

- game.ts: 1,562 -> 1,155 lines
- Game keeps thin delegates for the public render API used by tests
- Drops the now-unused GameConfig import from game.ts
